re PR bootstrap/53912 (bootstrap fails using default c++ mode in stage 2 and 3 for native x86_64-w64-mingw32)
	PR target/53912
	* stmt.c (compute_cases_per_edge): Cast from pointer via intptr_t.
	(expand_case): Likewise.

From-SVN: r194350

+++ gcc/stmt.c
@@ -2061,7 +2061,7 @@ compute_cases_per_edge (gimple stmt)
       tree lab = CASE_LABEL (elt);
       basic_block case_bb = label_to_block_fn (cfun, lab);
       edge case_edge = find_edge (bb, case_bb);
-      case_edge->aux = (void *)((long)(case_edge->aux) + 1);
+      case_edge->aux = (void *)((intptr_t)(case_edge->aux) + 1);
     }
 }
 
@@ -2176,7 +2176,7 @@ expand_case (gimple stmt)
       edge case_edge = find_edge (bb, case_bb);
       case_list = add_case_node (
           case_list, low, high, lab,
-          case_edge->probability / (long)(case_edge->aux),
+          case_edge->probability / (intptr_t)(case_edge->aux),
           case_node_pool);
     }
   pointer_set_destroy (seen_labels);
